pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bilinote-cli"
version = "0.1.0"
description = "AI 视频笔记生成 CLI 工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.5",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.13.0",
    "attrs>=26.1.0",
    "av>=17.0.0",
    "bilibili-api-python>=17.4.1",
    "certifi>=2026.2.25",
    "charset-normalizer>=3.4.7",
    "ctranslate2>=4.7.1",
    "distro>=1.9.0",
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "filelock>=3.28.0",
    "flatbuffers>=25.12.19",
    "frozenlist>=1.8.0",
    "fsspec>=2026.3.0",
    "future>=1.0.0",
    "gmssl>=3.2.2",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface-hub>=1.11.0",
    "idna>=3.11",
    "ipykernel>=7.2.0",
    "jiter>=0.14.0",
    "keyring>=25.7.0",
    "mlx-whisper>=0.4.3",
    "modelscope>=1.35.4",
    "mpmath>=1.3.0",
    "multidict>=6.7.1",
    "numpy>=2.4.4",
    "onnxruntime>=1.24.4",
    "openai>=2.32.0",
    "packaging>=26.1",
    "pillow>=12.2.0",
    "propcache>=0.4.1",
    "protobuf>=7.34.1",
    "pydantic>=2.13.2",
    "pydantic-core>=2.46.2",
    "pytest>=9.0.3",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "setuptools>=82.0.1",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sympy>=1.14.0",
    "tenacity>=9.1.4",
    "tokenizers>=0.22.2",
    "tqdm>=4.67.3",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "urllib3>=2.6.3",
    "yarl>=1.23.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.17",
]

[project.scripts]
bilinote = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
